requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pywin32==306
Requests==2.32.3
requests_html==0.10.0
